Break ConnectorFiltering class to separate files, add JavaDocs, add missing BWC tests, add requested changes to connector.list YAML test, put endpoints behind a feature flag
