+++ rest-api-spec/src/main/resources/rest-api-spec/api/connector.get.json
@@ -2,10 +2,11 @@
   "connector.get": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/enterprise-search/current/connectors.html",
-      "description": "Returns a connector document."
+      "description": "Returns the details about a connector."
     },
     "stability": "experimental",
-    "visibility": "public",
+    "visibility": "feature_flag",
+    "feature_flag": "es.connector_api_feature_flag_enabled",
     "headers": {
       "accept": [
         "application/json"
