+++ x-pack/plugin/ent-search/build.gradle
@@ -14,8 +14,7 @@ base {
 }
 
 dependencies {
-  testImplementation project(path: ':x-pack:plugin:eql')
-    compileOnly project(path: xpackModule('core'))
+  compileOnly project(path: xpackModule('core'))
   implementation project(xpackModule('search-business-rules'))
   api project(':modules:lang-mustache')
